requirements-linux.txt
baidu-aip>=2.2.18.0
certifi>=2020.6.20
cffi>=1.14.0
chardet>=3.0.4
configobj>=5.0.6
idna>=2.9
Pillow>=7.1.2
pycparser>=2.20
PyQt5>=5.15.0
PyQt5-sip>=12.8.0
requests>=2.24.0
six>=1.15.0
system-hotkey>=1.0.3
urllib3>=1.25.9
xcffib>=0.9.0
xpybutil>=0.0.6
